+++ docs/CHANGELOG.md
@@ -1,6 +1,13 @@
 # GTC :: Namespace - CHANGELOG
 
-## [1.0.0] - 2022-07-13
+## [0.1.2] - 2022-07-14
+* **[ref]** ```Object#namespace``` to provide self as 'klass'
+* **[ref]** GTC::Namespace::Base#initialize to provide any object & auto-detect it's class
+* **[ref]** docs
+* **[ref]** gemspec
+* **[fix]** specs
+
+## [0.1.1] - 2022-07-13
 * **[add]** docs
 * **[ref]** cleanup